refactor: extract admin settings tabs
